requirements.txt
aiogram==3.4.1
pydantic==2.5.3
aiohttp~=3.9.0
python-dotenv
aiofiles
aiosqlite
